fix(campaigns): unify result types and move interface
Address review comments from @MRashad26 on PR #895:

- campaign.interface.ts: add CampaignResult discriminated union
  (CampaignCreateResult | LinkedInCampaignCreateResult), require
  platform: 'google-ads' on CampaignCreateResult, update
  CampaignCreateResponse.campaigns to CampaignResult[]
- campaign.interface.ts: extract CampaignCreateBase for shared fields
  between CampaignCreateRequest and LinkedInCampaignCreateRequest;
  both now extend CampaignCreateBase
- campaign.interface.ts: move LinkedInTargetingProfileConfig from
  constants file to interfaces file; barrel re-exports automatically
- campaign.constants.ts: re-add disabled: true on linkedin-ads until
  backend dispatch PR lands; import LinkedInTargetingProfileConfig
- campaign-proxy.service.ts: add platform: 'google-ads' to Google Ads
  result returns; use CampaignResult[] for results accumulator
- implementation-tab.component.ts: update results signal to
  CampaignResult[] to accept both platform shapes

LFXV2-2023
